+++ examples/data-sources/ispconfig_cron_task/data-source.tf
@@ -1,0 +1,3 @@
+data "ispconfig_cron_task" "example" {
+  id = 1
+}